requirements/install.txt
click
gitpython
toolspy
ipython